+++ app/src/ai/blocklist/input_model.rs
@@ -380,9 +380,12 @@ impl BlocklistAIInputModel {
     pub fn is_terminal_use_active_or_pending(&self) -> bool {
         let model = self.model.lock();
         let active_block = model.block_list().active_block();
+        // Keep AI input locked while an agent-requested command is waiting for its
+        // CLI subagent, while the user has tagged the agent in, or while the user
+        // can hand control of an active monitored command back to the agent.
         active_block.is_agent_driving_command()
             || active_block.is_agent_tagged_in()
-            || active_block.long_running_control_state().is_some()
+            || active_block.is_eligible_for_agent_handoff()
     }
     pub fn last_ai_autodetection_source(&self) -> Option<InputTypeAutoDetectionSource> {
         self.last_ai_autodetection_source
