+++ crates/warp_tui/src/terminal_session_view_tests.rs
@@ -1,200 +1,11 @@
-use std::cell::RefCell;
-use std::collections::HashMap;
-use std::rc::Rc;
-use std::sync::Arc;
+use warp::tui_export::{export_conversation_markdown, PtyIntent, PtyIntentEvent};
 
-use parking_lot::FairMutex;
-use warp::tui_export::{
-    export_conversation_markdown, AIAgentActionId, AIConversationId, AgentInteractionMetadata,
-    BlockId, LongRunningCommandControlState, PtyIntent, PtyIntentEvent, TaskId, TerminalModel,
-    TranscriptScope, UserTakeOverReason,
-};
-use warpui::EntityIdMap;
-use warpui_core::elements::tui::{TuiLayoutContext, TuiViewportWindow, TuiViewportedElement};
-use warpui_core::App;
-
-use super::{
-    export_file_success_message, hide_agent_requested_command_from_top_level,
-    raw_prompt_if_not_blank, routes_submission_to_terminal_use,
-    terminal_use_conversation_to_resume, terminal_use_interrupt_action, user_controlled_line_bytes,
-    TerminalUseInterruptAction, TuiTerminalSessionEvent,
-};
-use crate::tui_block_list_viewport_source::TuiBlockListViewportSource;
-
-fn model_with_finished_block(command: &str) -> (TerminalModel, BlockId) {
-    let mut model = TerminalModel::mock(None, None);
-    model
-        .block_list_mut()
-        .set_transcript_scope(TranscriptScope::Unfiltered);
-    model.simulate_block(command, "output\r\n");
-    let block_id = model
-        .block_list()
-        .blocks()
-        .iter()
-        .rev()
-        .find(|block| block.finished())
-        .expect("simulated block should exist")
-        .id()
-        .clone();
-    (model, block_id)
-}
+use super::{export_file_success_message, raw_prompt_if_not_blank, TuiTerminalSessionEvent};
 
 #[test]
 fn interrupt_event_projects_to_high_level_pty_intent() {
     let event = TuiTerminalSessionEvent::InterruptPty;
     assert!(matches!(event.pty_intent(), Some(PtyIntent::Interrupt)));
-}
-
-#[test]
-fn user_controlled_line_has_no_agent_cursor_movement_prefix() {
-    #[cfg(target_os = "windows")]
-    assert_eq!(user_controlled_line_bytes("hello"), b"hello\r");
-    #[cfg(not(target_os = "windows"))]
-    assert_eq!(user_controlled_line_bytes("hello"), b"hello\n");
-}
-
-fn mark_visible_agent_requested_command(
-    model: &mut TerminalModel,
-    block_id: &BlockId,
-    action_id: &AIAgentActionId,
-) {
-    model
-        .block_list_mut()
-        .mut_block_from_id(block_id)
-        .expect("block should exist")
-        .set_agent_interaction_mode(AgentInteractionMetadata::new(
-            Some(action_id.clone()),
-            AIConversationId::new(),
-            None,
-            None,
-            false,
-            false,
-        ));
-    model
-        .block_list_mut()
-        .set_visibility_of_block_for_ai_action(action_id, true);
-}
-
-#[test]
-fn spawned_agent_requested_command_has_zero_top_level_height() {
-    let action_id: AIAgentActionId = "action".to_owned().into();
-    let (mut model, block_id) = model_with_finished_block("cargo build");
-    mark_visible_agent_requested_command(&mut model, &block_id, &action_id);
-    let model = Arc::new(FairMutex::new(model));
-
-    {
-        let model = model.lock();
-        let block_list = model.block_list();
-        let block = block_list
-            .block_with_id(&block_id)
-            .expect("block should exist");
-        assert!(block.is_visible(block_list.transcript_scope()));
-        assert!(block_list.block_heights().summary().height.as_f64() > 0.0);
-    }
-
-    assert!(hide_agent_requested_command_from_top_level(
-        &model,
-        Some(&action_id),
-    ));
-
-    let model = model.lock();
-    let block_list = model.block_list();
-    let block = block_list
-        .block_with_id(&block_id)
-        .expect("block should exist");
-    assert!(!block.is_visible(block_list.transcript_scope()));
-    assert_eq!(block_list.block_heights().summary().height.as_f64(), 0.0);
-}
-
-#[test]
-fn spawned_user_command_keeps_its_top_level_height() {
-    let (model, block_id) = model_with_finished_block("sleep 10");
-    let model = Arc::new(FairMutex::new(model));
-    let height_before = model
-        .lock()
-        .block_list()
-        .block_heights()
-        .summary()
-        .height
-        .as_f64();
-
-    assert!(!hide_agent_requested_command_from_top_level(&model, None));
-
-    let model = model.lock();
-    let block_list = model.block_list();
-    let block = block_list
-        .block_with_id(&block_id)
-        .expect("block should exist");
-    assert!(block.is_visible(block_list.transcript_scope()));
-    assert_eq!(
-        block_list.block_heights().summary().height.as_f64(),
-        height_before
-    );
-}
-
-#[test]
-fn hidden_agent_requested_command_leaves_no_viewport_gap() {
-    App::test((), |app| async move {
-        app.read(|app| {
-            let action_id: AIAgentActionId = "action".to_owned().into();
-            let mut model = TerminalModel::mock(None, None);
-            model.simulate_block("cargo build", "agent output\r\n");
-            model.simulate_block("echo done", "done\r\n");
-            let agent_block_id = model
-                .block_list()
-                .blocks()
-                .iter()
-                .find(|block| block.command_to_string().contains("cargo build"))
-                .expect("agent command block should exist")
-                .id()
-                .clone();
-            let user_block_id = model
-                .block_list()
-                .blocks()
-                .iter()
-                .find(|block| block.command_to_string().contains("echo done"))
-                .expect("user command block should exist")
-                .id()
-                .clone();
-            mark_visible_agent_requested_command(&mut model, &agent_block_id, &action_id);
-            let model = Arc::new(FairMutex::new(model));
-
-            assert!(hide_agent_requested_command_from_top_level(
-                &model,
-                Some(&action_id),
-            ));
-
-            let expected_height = {
-                let model = model.lock();
-                let block_list = model.block_list();
-                block_list
-                    .block_with_id(&user_block_id)
-                    .expect("user block should exist")
-                    .height(block_list.transcript_scope())
-                    .as_f64()
-                    .ceil() as usize
-            };
-            let source =
-                TuiBlockListViewportSource::new(model, Rc::new(RefCell::new(HashMap::new())));
-            let mut rendered_views = EntityIdMap::default();
-            let mut layout_ctx = TuiLayoutContext {
-                rendered_views: &mut rendered_views,
-            };
-            let content = source.visible_items(
-                TuiViewportWindow {
-                    scroll_top: 0,
-                    viewport_height: u16::MAX,
-                },
-                80,
-                &mut layout_ctx,
-                app,
-            );
-
-            assert_eq!(content.content_height, expected_height);
-            assert_eq!(content.items.len(), 1);
-            assert_eq!(content.items[0].origin_y, 0);
-        });
-    });
 }
 
 #[test]
@@ -222,94 +33,4 @@ fn file_export_success_message_includes_destination_path() {
         export_file_success_message(&export),
         format!("Conversation exported to {}", export.path().display())
     );
-}
-
-#[test]
-fn only_agent_control_routes_submissions_to_terminal_use() {
-    let agent = LongRunningCommandControlState::Agent {
-        is_blocked: false,
-        should_hide_responses: false,
-    };
-    let user = LongRunningCommandControlState::User {
-        reason: UserTakeOverReason::Manual,
-    };
-
-    assert!(routes_submission_to_terminal_use(&agent));
-    assert!(!routes_submission_to_terminal_use(&user));
-}
-
-#[test]
-fn terminal_use_interrupt_follows_takeover_then_process_interrupt_policy() {
-    let agent = LongRunningCommandControlState::Agent {
-        is_blocked: false,
-        should_hide_responses: false,
-    };
-    assert_eq!(
-        terminal_use_interrupt_action(&agent),
-        TerminalUseInterruptAction::TakeControl
-    );
-
-    let stopped = LongRunningCommandControlState::User {
-        reason: UserTakeOverReason::Stop {
-            should_auto_resume: true,
-        },
-    };
-    assert_eq!(
-        terminal_use_interrupt_action(&stopped),
-        TerminalUseInterruptAction::InterruptCommand
-    );
-
-    let manual = LongRunningCommandControlState::User {
-        reason: UserTakeOverReason::Manual,
-    };
-    assert_eq!(
-        terminal_use_interrupt_action(&manual),
-        TerminalUseInterruptAction::InterruptCommand
-    );
-
-    let transferred = LongRunningCommandControlState::User {
-        reason: UserTakeOverReason::TransferFromAgent {
-            reason: "enter password".to_owned(),
-        },
-    };
-    assert_eq!(
-        terminal_use_interrupt_action(&transferred),
-        TerminalUseInterruptAction::InterruptCommand
-    );
-}
-
-#[test]
-fn completed_user_controlled_requested_command_resumes_unless_tearing_down() {
-    let mut model = TerminalModel::mock(None, None);
-    model.simulate_long_running_block("cargo build", "Compiling");
-    let conversation_id = AIConversationId::new();
-    let task_id = TaskId::new("terminal-use-task".to_owned());
-    let block_id = {
-        let block = model.block_list_mut().active_block_mut();
-        block.set_agent_interaction_mode_for_requested_command(
-            AIAgentActionId::from("requested-command".to_owned()),
-            Some(task_id.clone()),
-            conversation_id,
-        );
-        block
-            .set_agent_interaction_mode_for_agent_monitored_command(&task_id, conversation_id)
-            .expect("command should become agent monitored");
-        block
-            .take_over_control_for_user(UserTakeOverReason::Stop {
-                should_auto_resume: true,
-            })
-            .expect("user takeover should succeed");
-        block.id().clone()
-    };
-
-    assert_eq!(
-        terminal_use_conversation_to_resume(&model, &block_id),
-        Some(conversation_id)
-    );
-
-    model
-        .block_list_mut()
-        .active_block_mut()
-        .set_user_control_for_teardown();
-    assert_eq!(terminal_use_conversation_to_resume(&model, &block_id), None);
 }
