+++ crates/warp_tui/src/transcript_view.rs
@@ -27,9 +27,7 @@ use super::terminal_block::should_render_terminal_block;
 use super::tui_block_list_viewport_source::{
     AgentBlockRegistry, CLISubagentBlockRegistry, TuiBlockListViewportSource,
 };
-use super::tui_cli_subagent_view::{
-    PendingTuiCLISubagentViews, TuiCLISubagentView, TuiCLISubagentViewEvent,
-};
+use super::tui_cli_subagent_view::{TuiCLISubagentView, TuiCLISubagentViewEvent};
 
 /// Rows of blank space above every transcript block. Terminal blocks get it
 /// via [`TRANSCRIPT_BLOCK_SPACING`]'s `padding_top`; agent blocks apply the
@@ -76,7 +74,6 @@ pub(super) struct TuiTranscriptView {
     model_events: ModelHandle<ModelEventDispatcher>,
     agent_blocks: AgentBlockRegistry,
     cli_subagent_blocks: CLISubagentBlockRegistry,
-    pending_cli_subagent_views: PendingTuiCLISubagentViews<ViewHandle<TuiCLISubagentView>>,
     viewport: TuiViewportedListState,
     selection: TuiSelectionHandle,
 }
@@ -102,7 +99,6 @@ impl TuiTranscriptView {
             model_events: model_events.clone(),
             agent_blocks: Rc::new(RefCell::new(HashMap::new())),
             cli_subagent_blocks: Rc::new(RefCell::new(HashMap::new())),
-            pending_cli_subagent_views: PendingTuiCLISubagentViews::default(),
             viewport: TuiViewportedListState::new_at_end(),
             selection: TuiSelectionHandle::default(),
         }
@@ -124,8 +120,6 @@ impl TuiTranscriptView {
     ) {
         if let Some(action_id) = action_id {
             let conversation_id = view.as_ref(ctx).conversation_id();
-            self.pending_cli_subagent_views
-                .remove(conversation_id, action_id);
             for agent_block in self.agent_blocks.borrow().values() {
                 if agent_block.as_ref(ctx).conversation_id() != conversation_id {
                     continue;
@@ -137,8 +131,9 @@ impl TuiTranscriptView {
                     return;
                 }
             }
-            self.pending_cli_subagent_views
-                .insert(conversation_id, action_id.clone(), view);
+            log::warn!(
+                "Unable to attach TUI CLI subagent for conversation {conversation_id:?} to action {action_id:?}"
+            );
             return;
         }
 
@@ -158,18 +153,12 @@ impl TuiTranscriptView {
 
     pub(super) fn detach_cli_subagent(
         &mut self,
-        conversation_id: AIConversationId,
         action_id: Option<&AIAgentActionId>,
         view_id: EntityId,
         ctx: &mut ViewContext<Self>,
     ) {
         if let Some(action_id) = action_id {
-            self.pending_cli_subagent_views
-                .remove(conversation_id, action_id);
             for agent_block in self.agent_blocks.borrow().values() {
-                if agent_block.as_ref(ctx).conversation_id() != conversation_id {
-                    continue;
-                }
                 let detached = agent_block.update(ctx, |agent_block, ctx| {
                     agent_block.set_cli_subagent_view(action_id, None, ctx)
                 });
@@ -353,12 +342,12 @@ impl TuiTranscriptView {
         let terminal_model = self.model.clone();
         let view = ctx.add_typed_action_tui_view(|ctx| {
             TuiAIBlock::new(
-                (conversation_id, exchange_id),
+                conversation_id,
+                exchange_id,
                 block_model,
                 action_model,
                 &model_events,
                 terminal_model,
-                self.pending_cli_subagent_views.clone(),
                 ctx,
             )
         });
@@ -501,8 +490,6 @@ impl TuiTranscriptView {
         conversation_id: AIConversationId,
         ctx: &mut ViewContext<Self>,
     ) {
-        self.pending_cli_subagent_views
-            .remove_conversation(conversation_id);
         let mut view_ids = self
             .agent_blocks
             .borrow()
@@ -557,7 +544,6 @@ impl TuiTranscriptView {
         view_ids.extend(self.cli_subagent_blocks.borrow().keys().copied());
         self.agent_blocks.borrow_mut().clear();
         self.cli_subagent_blocks.borrow_mut().clear();
-        self.pending_cli_subagent_views.clear();
         self.selection.clear();
         let mut model = self.model.lock();
         for view_id in view_ids {
