+++ crates/warp_tui/src/transcript_view_tests.rs
@@ -500,7 +500,8 @@ fn insert_test_agent_block(
         let terminal_model = view.model.clone();
         let agent_block = ctx.add_typed_action_tui_view(|ctx| {
             TuiAIBlock::new(
-                (conversation_id, exchange_id),
+                conversation_id,
+                exchange_id,
                 Rc::new(FakeAgentBlockModel {
                     inputs,
                     status: AIBlockOutputStatus::Pending,
@@ -508,7 +509,6 @@ fn insert_test_agent_block(
                 action_model,
                 &model_events,
                 terminal_model,
-                Default::default(),
                 ctx,
             )
         });
@@ -618,7 +618,8 @@ fn append_test_agent_block(
     let terminal_model = view.model.clone();
     let agent_block = ctx.add_tui_view(|ctx| {
         TuiAIBlock::new(
-            (conversation_id, exchange_id),
+            conversation_id,
+            exchange_id,
             Rc::new(FakeAgentBlockModel {
                 inputs: Vec::new(),
                 status,
@@ -626,7 +627,6 @@ fn append_test_agent_block(
             action_model,
             &model_events,
             terminal_model,
-            Default::default(),
             ctx,
         )
     });
